+++ ntex/tests/http_openssl.rs
@@ -439,3 +439,21 @@ async fn test_h2_on_connect() {
     let response = srv.srequest(Method::GET, "/").send().await.unwrap();
     assert!(response.status().is_success());
 }
+
+#[ntex::test]
+async fn test_ssl_handshake_timeout() {
+    use std::io::Read;
+
+    let srv = test_server(move || {
+        HttpService::build()
+            .ssl_handshake_timeout(50)
+            .h2(|_| ok::<_, io::Error>(Response::Ok().finish()))
+            .openssl(ssl_acceptor())
+            .map_err(|_| ())
+    });
+
+    let mut stream = std::net::TcpStream::connect(srv.addr()).unwrap();
+    let mut data = String::new();
+    let _ = stream.read_to_string(&mut data);
+    assert!(data.is_empty());
+}